+++ sanyan/__init__.py
@@ -13,4 +13,4 @@
 
 from __future__ import annotations
 
-__version__ = '3.58.0'
+__version__ = '3.59.0'